requirements.txt
alembic==1.14.1
bcrypt==4.2.1
blinker==1.9.0
click==8.1.8
coverage==7.6.12
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flask-sqlalchemy-stubs==0.2
greenlet==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
mypy==1.14.1
mypy-extensions==1.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
pytest-cov==6.0.0
python-dotenv==1.0.1
ruff==0.9.4
SQLAlchemy==2.0.37
types-click==7.1.8
types-Flask==1.1.6
types-Jinja2==2.11.9
types-MarkupSafe==1.1.10
types-Werkzeug==1.0.9
typing_extensions==4.12.2
Werkzeug==3.1.3
